requirements/common.txt
django~=3.1.7
django-bootstrap4
pillow
python-dotenv
requests
